rock paper scissors
The purpose of this code is to receive an input number from the user. this number will be used to search the index of options within the list. the cpu generates a random number within 0 and 2, and this number has the same function as the user's number. At the end, the program asks the user whether he wants to play more or not